AI_recommendation/Python_API/requirements.txt
flask
flask-cors
sys
os
numpy
pandas
scikit-learn
python-dotenv